moved protos to comms-broker package + starting to use own types in comms index instead of protobuf
