+++ kernel/packages/shared/comms-broker/BrokerConnection.ts
@@ -13,7 +13,7 @@ import { Auth } from 'shared/auth/Auth'
 import { AuthData } from './proto/comms'
 import { Message } from 'google-protobuf'
 import { commConfigurations } from 'config'
-import { Stats } from './debug'
+import { Stats } from '../comms/debug'
 import { IBrokerConnection, BrokerMessage, SocketReadyState } from './IBrokerConnection'
 import { ILogger, createLogger } from 'shared/logger'
 import { Observable } from '../../decentraland-ecs/src/ecs/Observable'
