+++ kernel/packages/shared/comms-interface/types.ts
@@ -91,3 +91,51 @@ export type Pose = [number, number, number, number, number, number, number]
 export type PoseInformation = {
   v: Pose
 }
+
+// message AuthData {
+//     string signature = 1;
+//     string identity = 2;
+//     string timestamp = 3;
+//     string access_token = 4;
+// }
+
+// enum Category {
+//      UNKNOWN = 0;
+//      POSITION = 1;
+//      PROFILE = 2;
+//      CHAT = 3;
+//      SCENE_MESSAGE = 4;
+// }
+
+// message DataHeader {
+//     Category category = 1;
+// }
+
+// message PositionData {
+//     Category category = 1;
+//     double time = 2;
+//     float position_x = 3;
+//     float position_y = 4;
+//     float position_z = 5;
+//     float rotation_x = 6;
+//     float rotation_y = 7;
+//     float rotation_z = 8;
+//     float rotation_w = 9;
+// }
+export type Package<T> = {
+  time: number
+  data: T
+}
+
+// message ProfileData {
+//     Category category = 1;
+//     double time = 2;
+//     string profile_version = 3;
+// }
+
+// message ChatData {
+//     Category category = 1;
+//     double time = 2;
+//     string message_id = 3;
+//     string text = 4;
+// }
