+++ kernel/packages/shared/comms-interface/utils.ts
